create Dashboard component
